tg_project/appturbomanager-backend-main/requirements.txt
aiohttp==3.11.16
aiogram==3.19.0
alembic
anthropic
asyncpg
bcrypt
boto3
cent
fastapi
groq
langchain
langchain-anthropic
langchain-openai
Pillow
passlib
pydantic
pydantic[email]
psycopg2-binary
PyJWT
python-multipart
redis
ruff>=0.3.0
SQLAlchemy-Utils
sqlalchemy
telethon
uvicorn
uvicorn[standard]
python-dotenv
python-jose
Werkzeug
qrcode[pil]